# some faster `Tuple` and `merge(::NamedTuple, ::Zip)` methods