Before:
```
julia> names = [:a,:b,:c,:d,:e,:f,:g,:h,:i]
julia> spl(n) = (n...,)
julia> tup(n) = Tuple(n)
julia> @btime spl(names)
  205.563 ns (1 allocation: 80 bytes)

julia> @btime tup(names)
  250.763 ns (2 allocations: 240 bytes)

julia> values = [1:9;]
julia> nt(k, v) = (; zip(k,v)...)
julia> @btime nt(names, values)
  1.798 μs (16 allocations: 1.44 KiB)
```
after:
```
julia> @btime tup(names)
  214.543 ns (1 allocation: 80 bytes)

julia> @btime nt(names, values)
  1.362 μs (6 allocations: 400 bytes)
```